+++ model.py
@@ -53,7 +53,9 @@ class TrainingModel(nn.Module):
                 return_dict=True,
             )
             hidden_states = base_outputs.last_hidden_state
-            logits = self.model.lm_head(hidden_states)
+
+        # Compute LM head in full precision for more stable log-prob/loss math.
+        logits = self.model.lm_head(hidden_states.to(torch.float32)).to(torch.float32)
 
         with torch.autocast(device_type="cuda", dtype=torch.bfloat16):
             value = self.value_head(self.value_norm(hidden_states)).squeeze(
